build: Skip 'clean' and 'clean-container' before docker image builds.
Running 'make clean' will not change 'go build' output, so it is not
necessary, especially now that building with 'DOCKER_BUILDKIT=1' works
and the '.dockerignore' files are automatically generated so that
binaries, temporary files, etc. are not tranferred to the docker build
container.

This will make the build output easier to read and save some build time, too.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>