iptables, loader: add rules for multi-node NodePort traffic on EKS
Multi-node NodePort traffic for EKS needs a set of specific rules that
are usually set by the aws daemonset:

    # sysctl -w net.ipv4.conf.eth0.rp_filter=2
    # iptables -t mangle -A PREROUTING -i eth0 -m comment --comment "AWS, primary ENI" -m addrtype --dst-type LOCAL --limit-iface-in -j CONNMARK --set-xmark 0x80/0x80
    # iptables -t mangle -A PREROUTING -i eni+ -m comment --comment "AWS, primary ENI" -j CONNMARK --restore-mark --nfmask 0x80 --ctmask 0x80
    # ip rule add fwmark 0x80/0x80 lookup main

These rules mark packets coming from another node through eth0, and
restore the mark on the return path to force a lookup into the main
routing table. Without them, the "ip rules" set by the cilium-cni plugin
tell the host to lookup into the table related to the VPC for which the
CIDR used by the endpoint has been configured.

We want to reproduce equivalent rules to ensure correct routing, or
multi-node NodePort traffic will not be routed correctly. This could be
observed with the pod-to-b-multi-node-nodeport pod from connectivity
check never getting ready.

This commit makes the loader and iptables module create the relevant
rules when IPAM is ENI and egress masquerading is in use. The rules are
nearly identical to those from the aws daemonset (different comments,
different interface prefix for conntrack return path, explicit
preference for ip rule):

    # sysctl -w net.ipv4.conf.<egressMasqueradeInterfaces>.rp_filter=2
    # iptables -t mangle -A PREROUTING -i <egressMasqueradeInterfaces> -m comment --comment "cilium: primary ENI" -m addrtype --dst-type LOCAL --limit-iface-in -j CONNMARK --set-xmark 0x80/0x80
    # iptables -t mangle -A PREROUTING -i lxc+ -m comment --comment "cilium: primary ENI" -j CONNMARK --restore-mark --nfmask 0x80 --ctmask 0x80
    # ip rule add fwmark 0x80/0x80 lookup main pref 109

Fixes: #12098
Signed-off-by: Quentin Monnet <quentin@isovalent.com>